refactor(private-rpc): extract subscription types
