+++ crates/rpc/src/handlers.rs
@@ -3,56 +3,17 @@
 //! Each handler calls the underlying EthApi via the [`ZoneRpcApi`] trait,
 //! which performs typed privacy redactions internally before serialization.
 
-use std::{future::Future, pin::Pin};
-
 use alloy_primitives::{Address, B256, Bytes};
 use alloy_rpc_types_eth::{BlockId, BlockNumberOrTag, Filter, FilterId, state::StateOverride};
-use futures::Stream;
 use serde_json::{Value, value::RawValue};
 use tempo_alloy::rpc::TempoTransactionRequest;
 use tracing::warn;
 
 use crate::{
     auth::AuthContext,
+    subscription::BoxWsSubscriptionFut,
     types::{BoxFut, JsonRpcError, JsonRpcRequest, JsonRpcResponse, MethodTier, classify_method},
 };
-
-/// A boxed stream of serialized websocket subscription items.
-pub type WsSubscriptionStream =
-    Pin<Box<dyn Stream<Item = Result<Box<RawValue>, JsonRpcError>> + Send + 'static>>;
-
-/// A boxed future that resolves to a websocket subscription.
-pub type BoxWsSubscriptionFut<'a> =
-    Pin<Box<dyn Future<Output = Result<WsSubscription, JsonRpcError>> + Send + 'a>>;
-
-/// A transport-agnostic websocket subscription returned by [`ZoneRpcApi`].
-pub struct WsSubscription {
-    /// Stream of serialized `eth_subscription` payloads.
-    pub stream: WsSubscriptionStream,
-    /// Optional upstream filter ID that must be cleaned up on unsubscribe.
-    pub upstream_filter_id: Option<FilterId>,
-}
-
-impl WsSubscription {
-    /// Create a subscription backed by a direct event stream.
-    pub fn new(stream: WsSubscriptionStream) -> Self {
-        Self {
-            stream,
-            upstream_filter_id: None,
-        }
-    }
-
-    /// Create a subscription that owns an upstream filter.
-    pub fn with_upstream_filter(
-        stream: WsSubscriptionStream,
-        upstream_filter_id: FilterId,
-    ) -> Self {
-        Self {
-            stream,
-            upstream_filter_id: Some(upstream_filter_id),
-        }
-    }
-}
 
 /// Interface to the underlying reth EthApi for the private zone RPC.
 ///
