+++ crates/rpc/src/lib.rs
@@ -14,6 +14,7 @@ pub mod policy;
 pub mod provider;
 pub mod proxy;
 pub mod server;
+pub mod subscription;
 pub mod types;
 mod ws;
